renamed servicetype package to service
